Refactor statediff out to be used elsewhere
