+++ utils/statediff/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "statediff"
+version = "0.1.0"
+authors = ["ChainSafe Systems <info@chainsafe.io>"]
+edition = "2018"
+
+[dependencies]
+serde_json = "1.0"
+blockstore = { package = "ipld_blockstore", path = "../../ipld/blockstore/", features = [
+    "resolve"
+] }
+cid = { package = "forest_cid", path = "../../ipld/cid", features = [
+    "cbor",
+    "json"
+] }
+difference = "2.0"
+colored = "2.0"
+ipld_hamt = { path = "../../ipld/hamt", features = ["ignore-dead-links"] }
+address = { package = "forest_address", path = "../../vm/address", features = [
+    "json"
+] }
+serde = { version = "1.0", features = ["derive"] }
+ipld = { package = "forest_ipld", path = "../../ipld", features = ["json"] }
+vm = { package = "forest_vm", path = "../../vm" }
+fil_types = { path = "../../types" }
